Change like unlike component path
